requirements.txt
flask
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
googleapiclient
pyusb
